+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Fix history server stderr/stdout. Now is possible to set log level through SPARK_LOG_LEVEL
 * Removed mesos security from History Server and unified environment variable VAULT_HOSTS
 * Secret folder path configurable
+* Changed log format according to Stratio standards
 
 ## 2.2.0.4 (January 11, 2018)